+++ minesweeper/src/modules/MainAlgorithm.ts
